backend/project/requirements.txt
fastapi==0.120.4
uvicorn==0.18.3
psycopg2-binary==2.9.10
pyjwt==2.8.0
pydantic
python-dotenv==1.0.1
requests==2.31.0
fhirclient==4.1.0
